requirement.txt
python-multipart
starlette
fastapi 
uvicorn 
pillow 
numpy 